pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ezquant"
version = "0.1.0"
description = "Policy-gated quantitative image analysis pipeline"
requires-python = ">=3.10"
dependencies = [
  "numpy",
  "pandas",
  "pyyaml",
  "tifffile",
  "scipy",
  "scikit-image",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
